requirements.txt
requests
playwright
python-dotenv
functions_framework